requirements.txt
# Manim 教育视频多智能体流水线 – 依赖
# 安装: pip install -r requirements.txt

# 代码生成与 API
openai>=1.30.0
python-dotenv>=1.0.1

# 评测管道：CV + 数值
opencv-python>=4.8.0
numpy>=1.24.0

# 渲染与 TTS
manim>=0.18.0
edge-tts>=6.1.0
mutagen>=1.47.0

# 可选：OCR（需系统安装 Tesseract）
# pytesseract>=0.3.10
